Eliminate useless Tty type
I want the name for renaming PlatformInput.
